+++ src/main/java/com/houseofpizza/repository/specification/builder/StatusSpecificationBuilder.java
@@ -10,7 +10,7 @@ import lombok.experimental.UtilityClass;
 @UtilityClass
 public class StatusSpecificationBuilder {
 
-    public Specification<Status> withStatusIdEqualTo(Integer statusId) {
+    public Specification<Status> withStatusIdEqualTo(Long statusId) {
         return Specification.where(StatusSpecification.withStatusIdEqualTo(statusId));
     }
